## 작업 내용
> 이번 PR에 포함될 작업 내용에 대한 간략한 설명

Authorize HTTP Request 대신 **Method Security** 방법으로 세분화된 권한을 부여하였다.
- MarketController, ArticleController -> 게시글 조회, 게시글리스트 조회에는 모든 접근을
허용하고, 수정 삭제의 경우 @preAuthorize로 role을 확인한 후, 메서드 호출자와 게시글 작성자가 동일한지 판별하는
로직을 서비스 계층에 추가하였다.
- ClubController -> 조회에는 모든 접근을 허용하고, 수정, 삭제의 경우 클럽 소유자를 판별하는 로직을 서비스
계층에 추가하였다.
- AdminController -> role이 Admin인지 확인하였다.
- UserController -> 로그인 여부만 확인하였다.

<br>

- [x] AdminController
- [x] ClubController
- [x] MarketController
- [x] ArticleController
- [x] UserController

## 참고(선택)
> 리뷰어가 참고할만한 내용

1. 설정 
Method Security는 Spring AOP기반으로 구축되어 있기 때문에, Method Security를 시작하려면
@Configuration이 있는 곳에 **@EnableMethodSecurity**를 추가한다.

2. 메서드 권한 부여
- **@PreAuthroize**
- @PostAuthorize
- @PreFilter
- @PostFilter



https://velog.io/@padomay1352/spring-security-method-%EA%B6%8C%ED%95%9CAhuthorization-%EC%84%A4%EC%A0%95

### #️⃣이슈 번호
> 연결된 이슈 번호 (#번호)
Resolves #143

+++ src/main/java/com/charmroom/charmroom/service/ClubService.java
@@ -102,9 +102,13 @@ public class ClubService {
     }
 
     @Transactional
-    public ClubDto update(Integer clubId, ClubDto clubDto) {
+    public ClubDto update(Integer clubId, ClubDto clubDto, String username) {
         Club club = clubRepository.findById(clubId).orElseThrow(() ->
                 new BusinessLogicException(BusinessLogicError.NOTFOUND_CLUB, "clubId: " + clubId));
+
+        if (!username.equals(club.getOwner().getUsername())) {
+            throw new BusinessLogicException(BusinessLogicError.UNAUTHORIZED_CLUB);
+        }
 
         club.updateName(clubDto.getName());
         club.updateDescription(clubDto.getDescription());
@@ -114,28 +118,40 @@ public class ClubService {
     }
 
     @Transactional
-    public ClubDto changeOwner(Integer clubId, String username) {
+    public ClubDto changeOwner(Integer clubId, String updatedName, String username) {
         Club club = clubRepository.findById(clubId).orElseThrow(() ->
                 new BusinessLogicException(BusinessLogicError.NOTFOUND_CLUB, "clubId: " + clubId));
 
-        User user = userRepository.findByUsername(username).orElseThrow(() ->
-                new BusinessLogicException(BusinessLogicError.NOTFOUND_USER, "username: " + username));
+        User user = userRepository.findByUsername(updatedName).orElseThrow(() ->
+                new BusinessLogicException(BusinessLogicError.NOTFOUND_USER, "username: " + updatedName));
+
+        if(!username.equals(club.getOwner().getUsername())) {
+            throw new BusinessLogicException(BusinessLogicError.UNAUTHORIZED_CLUB);
+        }
 
         club.updateOwner(user);
         return ClubMapper.toDto(club);
     }
 
-    public void deleteClub(Integer clubId) {
+    public void deleteClub(Integer clubId, String username) {
         Club club = clubRepository.findById(clubId).orElseThrow(() ->
                 new BusinessLogicException(BusinessLogicError.NOTFOUND_CLUB, "clubId: " + clubId));
+
+        if (!username.equals(club.getOwner().getUsername())) {
+            throw new BusinessLogicException(BusinessLogicError.UNAUTHORIZED_CLUB);
+        }
 
         clubRepository.delete(club);
     }
 
     @Transactional
-    public ClubDto setImage(Integer clubId, MultipartFile imageFile) {
+    public ClubDto setImage(Integer clubId, MultipartFile imageFile, String username) {
         Club club = clubRepository.findById(clubId).orElseThrow(() ->
                 new BusinessLogicException(BusinessLogicError.NOTFOUND_CLUB, "clubId: " + clubId));
+
+        if (!username.equals(club.getOwner().getUsername())) {
+            throw new BusinessLogicException(BusinessLogicError.UNAUTHORIZED_CLUB);
+        }
 
         if (club.getImage() != null) {
             uploadUtil.deleteFile(club.getImage());
